+++ CHANGELOG.md
@@ -50,6 +50,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 - In Gitlab CI, upload junit reports for corona and lassen.
 
+- Initial support for IPC Shared Memory via a "SHARED" resource allocator.
+
+- Allocator::getStrategyName() to get name of the strategy used.
+
 ### Changed
 
 - Organized configuration options in config.hpp.in in alphabetical order.
@@ -109,6 +113,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
   
 - Fixed ranges used in the vendor allocator benchmark when HIP is enabled given 
   that hipMalloc allocates on 4k aligned pages.
+
+- Fixed broken allocation test with DEVICE_CONST memory
+
+- Fixed compile error in DynamicSizePool with CUDA 11 and C++17
 
 ## [v5.0.1] - 2021-03-31